planck: Adjust raise/compose behaviour
HOLD_ON_OTHER_KEY_PRESS_PER_KEY is a more correct behaviour for the raise/compose key:
if a key is pressed while raise/compose is held, it cannot be part of a compose sequence,
as compose must be released before starting the sequence. Thus this means we want the
raise/hold behaviour when this happens, which is what HOLD_ON_OTHER_KEY_PRESS_PER_KEY
provides.