More Ergo-L work
Not everything in Ergo-L will be easily supportable in QMK, in particular the
accents dead keys on RAlt+Shift.
I'll go with a modified QWERTY layout instead, and use the OS-level keymap instead.
The keyboard RAISE key will be mapped to RAlt, but when LOWER is pressed, it
will change to RAISE so that it's possible to access the ADJUST layer.