planck: Add ergol layer
When the OS uses an ergo-l keymap, the symbols layer no longer inputs
the expected characters for some of the keys.
It's easier to switch back to AltGR instead of RAISE for the key to the
right of space for now. This will use the symbol layer defined in the OS
ergo-L keymap.