planck: Start refining CapsWord
Enable CapsWord with double tap on shift, and add caps_word_press_user
implementation. For now it is the default one, but it will need to be
customized for ergo-l: since I use the OS-level layout, and QWERTY as the
physical keyboard layout, some keys interrupt CapsWord because they are symbols
in QWERTY but they correspond to characters in ergo-l. QWERTY / ergo-l k or
QWERTY , ergo-l h are such keys.